fix(tui): peek UX polish — shared colorization, follow fix, auto-refresh (#1840)
- Extract colorizeOutputLine/hasAnsiCodes/isPeekHeader to shared utils/outputColors.tsx
- AgentPeekPanel uses colorizeOutputLine for semantic coloring instead of blanket dimColor
- Fix Live tab follow indicator: check isFollowing state instead of scrollOffset === 0
- Add 2s auto-refresh polling to peek panel (was static snapshot)
- AgentsView uses shared isPeekHeader instead of inline regex
