+++ tui/src/views/agents/AgentPeekPanel.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import { Box, Text, useStdout } from 'ink';
 import type { Agent } from '../../types';
-
-// eslint-disable-next-line no-control-regex
-const ANSI_REGEX = /\x1b\[[0-9;]*m/;
+import { colorizeOutputLine } from '../../utils';
 
 export interface AgentPeekPanelProps {
   agent: Agent;
@@ -50,13 +48,11 @@ export function AgentPeekPanel({
         <Text dimColor>Loading...</Text>
       ) : (
         <Box flexDirection="column" width={contentWidth}>
-          {output.map((line, idx) =>
-            ANSI_REGEX.test(line) ? (
-              <Text key={idx} wrap="wrap">{line}</Text>
-            ) : (
-              <Text key={idx} wrap="wrap" dimColor>{line}</Text>
-            )
-          )}
+          {output.map((line, idx) => (
+            <Box key={idx}>
+              <Text wrap="wrap">{colorizeOutputLine(line)}</Text>
+            </Box>
+          ))}
         </Box>
       )}
     </Box>
